恢复部分丢失代码

+++ WangShunManager/Dtos/ProductDto.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string ProductName { get; set; }
         public int CategoryId { get; set; }
-        public int ParValue { get; set; }
+        public decimal ParValue { get; set; }
         public int State { get; set; }
     }
 }
